Color, Image info, code correction
Fixing the colors of the website to make sure it has a good contrast. Image info has also been added to the images aswell as lowering the size of images. Lastly has the code been reviewed by a validator and the  code errors such as bad syntax have been fixed. An example is letting links act as buttons and switching from article to div of image.